moved middleware transformers and added new ones
