+++ Tests/HttpPipelineTests/HttpPipelineTests.swift
@@ -8,6 +8,10 @@ import SnapshotTesting
 private let conn = connection(from: URLRequest(url: URL(string: "/")!))
 
 class HttpPipelineTests: XCTestCase {
+  override func setUp() {
+    record = true
+  }
+  
   func testPipeline() {
     let middleware: Middleware<StatusLineOpen, ResponseEnded, Prelude.Unit, Data?> =
       writeStatus(.ok)
@@ -20,40 +24,6 @@ class HttpPipelineTests: XCTestCase {
     let middleware: Middleware<StatusLineOpen, ResponseEnded, Prelude.Unit, Data?> =
       writeStatus(.ok)
         >>> respond(html: "<p>Hello, world</p>")
-
-    assertSnapshot(matching: middleware(conn))
-  }
-
-  func testRedirect() {
-    let middleware: Middleware<StatusLineOpen, ResponseEnded, Prelude.Unit, Data?> = redirect(to: "/sign-in")
-
-    assertSnapshot(matching: middleware(conn))
-  }
-
-  func testRedirect_AdditionalHeaders() {
-    let middleware: Middleware<StatusLineOpen, ResponseEnded, Prelude.Unit, Data?> =
-      redirect(to: "/sign-in", headersMiddleware: writeHeader("Pass-through", "hello!"))
-
-    assertSnapshot(matching: middleware(conn))
-  }
-
-  func testBasicAuth_Unauthorized() {
-    let middleware: Middleware<StatusLineOpen, ResponseEnded, Prelude.Unit, Data?> =
-      basicAuth(user: "Hello", password: "World")
-        <| writeStatus(.ok) >>> respond(html: "<p>Hello, world</p>")
-
-    assertSnapshot(matching: middleware(conn))
-  }
-
-  func testBasicAuth_Authorized() {
-    let middleware: Middleware<StatusLineOpen, ResponseEnded, Prelude.Unit, Data?> =
-      basicAuth(user: "Hello", password: "World")
-        <| writeStatus(.ok) >>> respond(html: "<p>Hello, world</p>")
-
-    let conn = connection(
-      from: URLRequest(url: URL(string: "/")!)
-        |> \.allHTTPHeaderFields .~ ["Authorization": "Basic SGVsbG86V29ybGQ="]
-    )
 
     assertSnapshot(matching: middleware(conn))
   }
@@ -76,19 +46,6 @@ class HttpPipelineTests: XCTestCase {
         >>> writeHeader(.setCookie(["user_id": "123456"]))
         >>> writeHeader(.setCookie(["lang": "es"]))
         >>> respond(html: "<p>Hello, world</p>")
-
-    assertSnapshot(matching: middleware(conn))
-  }
-
-  func testContentLengthMiddlewareTransformer() {
-    let middleware: Middleware<StatusLineOpen, ResponseEnded, Prelude.Unit, Data?> =
-      contentLength
-        <| writeStatus(.ok)
-        >>> writeHeader(.contentType(.html))
-        >>> closeHeaders
-        >>> map(const(Data()))
-        >>> send("<p>Hello, world</p>".data(using: .utf8))
-        >>> end
 
     assertSnapshot(matching: middleware(conn))
   }
